fix: silence MSVC deprecation warnings and ensure float type for layout dimensions

+++ src/ui.c
@@ -680,7 +680,7 @@ void ui_render(struct nk_context *nk, app_state_t *state, int width, int height)
     }
 
     if (nk_begin(nk, "Wasteland",
-                 nk_rect(0, 0, width, height),
+                 nk_rect(0, 0, (float)width, (float)height),
                  NK_WINDOW_NO_SCROLLBAR | NK_WINDOW_BACKGROUND))
     {
         /* ========================= HEADER ========================= */
@@ -722,9 +722,9 @@ void ui_render(struct nk_context *nk, app_state_t *state, int width, int height)
 
         /* =================== MAIN SPLIT AREA =================== */
         if (state->left_panel_collapsed) {
-            nk_layout_row_dynamic(nk, height - 70, 1);
+            nk_layout_row_dynamic(nk, (float)(height - 70), 1);
         } else {
-            nk_layout_row_dynamic(nk, height - 70, 2);
+            nk_layout_row_dynamic(nk, (float)(height - 70), 2);
         }
 
         /* --------------------- LEFT PANEL --------------------- */
